+++ server/api/src/main/java/com/walmartlabs/concord/server/api/PerformedActionType.java
@@ -1,0 +1,7 @@
+package com.walmartlabs.concord.server.api;
+
+public enum PerformedActionType {
+
+    CREATED,
+    UPDATED
+}